xcopy-populator: Add destroyVM cmd to the compliance tool
